+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.1.1
 
 * Add support for detached branch #1
+* Add dirty flag to check for working tree status #2
 
 ### v0.1.0 (2019-11-27)